Add files via upload
It appears to work (all populations learn) but the results seem to not be as expected (hermaphrodite diploid is almost always winning, haploid is always winning).

Also, there is a significant difference in male vs female fitness due to females more often having dominant phenotypes. Imbalanced sex reproduction was not fully implemented as in the python file (top X% of either, leading to more females than males reproducing) because of this (hard difference, top X% of males but top Y% of females was).